+++ packages/svelte-maplibre-storymap/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @undp-data/svelte-maplibre-storymap
 
+## 0.3.21
+
+### Patch Changes
+
+- Updated dependencies [a331de4]
+  - @undp-data/svelte-undp-components@1.4.0
+
 ## 0.3.20
 
 ### Patch Changes